fix: growth-analysis/#429

+++ src/main/java/com/process/clash/application/compete/my/data/DataPoint.java
@@ -4,6 +4,6 @@ import java.time.LocalDate;
 
 public record DataPoint(
         LocalDate date,
-        Double rate
+        Long rate
 ) {
 }
